EventProcessMenu   -> EventMenuProcess EventCustomizeMenu -> EventMenu, Action: eCustomize
git-svn-id: file:///.data/svn_convert/svn@7260 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/_core/commands.h
@@ -56,7 +56,7 @@ protected:
     SIM::CommandsDef *getDef(unsigned id);
     CMenu *processMenu(unsigned id, void *param, int key);
     void customize(SIM::CommandsDef *def);
-    void customizeMenu(unsigned id);
+    void customizeMenu(unsigned long id);
     unsigned cur_id;
     CMDS_MAP bars;
     MENU_MAP menues;
